Refactoring + full beta reduction
Switch to menhir instead of ocamlyacc.
The code is now sliced in few modules.

Full beta reduction is implemented:
To know what is the name of the bounded variable for the premise, a
mechanism of "parametrized rules" is implemented, for the case of full
reduction, the name of the expected bounded variable is given regardless
of the name of the variable in the conclusion judgement, it nicely
matches with the nameless representation philosophy.
see the test file for an example.